backend/pyproject.toml
[tool.poetry]
name = "selfpr-backend"
version = "0.1.0"
description = "Self-introduction AI Agent Backend"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.12"

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"

# Cache / Queue
redis = {extras = ["hiredis"], version = "^5.2.0"}
celery = {extras = ["redis"], version = "^5.4.0"}

# AI
langchain = "^0.3.0"
langgraph = "^0.2.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.3.0"

# File Processing
pymupdf = "^1.24.0"
openpyxl = "^3.1.5"
python-docx = "^1.1.2"
Pillow = "^11.0.0"
# paddleocr = "^2.8.0"  # 별도 설치 (용량 큼)

# Document Export
weasyprint = "^62.3"

# Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^43.0.0"

# HTTP Client (웹 검색/크롤링)
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"

# Validation / Settings
pydantic = {extras = ["email"], version = "^2.10.0"}
pydantic-settings = "^2.6.0"

# Rate Limiting
slowapi = "^0.1.9"

# Utils
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.0"
faker = "^30.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
